Refactoring for new EntityManagerFactory initialization
