+++ src/AppserverIo/Appserver/Core/Api/Node/QueryCacheConfigurationNode.php
@@ -21,7 +21,7 @@
 namespace AppserverIo\Appserver\Core\Api\Node;
 
 use AppserverIo\Description\Api\Node\AbstractNode;
-use AppserverIo\Appserver\PersistenceContainer\Doctrine\CacheFactories\CacheConfigurationNodeInterface;
+use AppserverIo\Appserver\PersistenceContainer\Doctrine\V2\CacheFactories\CacheConfigurationNodeInterface;
 
 /**
  * DTO to transfer a query cache configuration.
@@ -55,7 +55,7 @@ class QueryCacheConfigurationNode extends AbstractNode implements CacheConfigura
      *
      * @param string $factory The factory class name to use
      */
-    public function __construct($factory = 'AppserverIo\Appserver\PersistenceContainer\Doctrine\CacheFactories\ArrayCacheFactory')
+    public function __construct($factory = 'AppserverIo\Appserver\PersistenceContainer\Doctrine\V2\CacheFactories\ArrayCacheFactory')
     {
         $this->factory = $factory;
     }
