--- src/AppserverIo/Appserver/PersistenceContainer/Doctrine/DriverFactories/DriverKeys.php
@@ -1,40 +1,0 @@
-<?php
-/**
- * AppserverIo\Appserver\PersistenceContainer\Doctrine\DriverFactories\DriverKeys
- *
- * NOTICE OF LICENSE
- *
- * This source file is subject to the Open Software License (OSL 3.0)
- * that is available through the world-wide-web at this URL:
- * http://opensource.org/licenses/osl-3.0.php
- *
- * PHP version 5
- *
- * @author    Tim Wagner <tw@appserver.io>
- * @copyright 2015 TechDivision GmbH <info@appserver.io>
- * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
- * @link      https://github.com/appserver-io/appserver
- * @link      http://www.appserver.io
- */
-
-namespace AppserverIo\Appserver\PersistenceContainer\Doctrine\DriverFactories;
-
-/**
- * Utility class with the driver keys.
- *
- * @author    Tim Wagner <tw@appserver.io>
- * @copyright 2015 TechDivision GmbH <info@appserver.io>
- * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
- * @link      https://github.com/appserver-io/appserver
- * @link      http://www.appserver.io
- */
-class DriverKeys
-{
-
-    /**
-     * The key with the value to use the simple annotation reader or not.
-     *
-     * @var string
-     */
-    const USE_SIMPLE_ANNOTATION_READER = 'useSimpleAnnotationReader';
-}
