+++ src/AppserverIo/Appserver/PersistenceContainer/Doctrine/V2/CacheFactories/XcacheCacheFactory.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * AppserverIo\Appserver\PersistenceContainer\Doctrine\CacheFactories\MemcacheCacheFactory
+ * AppserverIo\Appserver\PersistenceContainer\Doctrine\V2\CacheFactories\XcacheCacheFactory
  *
  * NOTICE OF LICENSE
  *
@@ -17,13 +17,12 @@
  * @link      https://github.com/appserver-io/rmi
  * @link      http://www.appserver.io
  */
+namespace AppserverIo\Appserver\PersistenceContainer\Doctrine\V2\CacheFactories;
 
-namespace AppserverIo\Appserver\PersistenceContainer\Doctrine\CacheFactories;
-
-use Doctrine\Common\Cache\RedisCache;
+use Doctrine\Common\Cache\XcacheCache;
 
 /**
- * The factory implementation for a RedisCache cache instance.
+ * The factory implementation for a XcacheCache cache instance.
  *
  * @author    Tim Wagner <tw@appserver.io>
  * @author    Bernhard Wick <bw@appserver.io>
@@ -32,7 +31,7 @@ use Doctrine\Common\Cache\RedisCache;
  * @link      https://github.com/appserver-io/rmi
  * @link      http://www.appserver.io
  */
-class MemcacheCacheFactory implements CacheFactoryInterface
+class XcacheCacheFactory implements CacheFactoryInterface
 {
 
     /**
@@ -44,12 +43,8 @@ class MemcacheCacheFactory implements CacheFactoryInterface
      */
     public static function get(array $configuration = array())
     {
-        if (extension_loaded('redis')) {
-            $redis = new \Redis();
-            $redis->connect($configuration[CacheKeys::HOST]);
-            $cache = new RedisCache();
-            $cache->setRedis($redis);
-            return $cache;
+        if (extension_loaded('xcache')) {
+            return new XcacheCache();
         }
     }
 }
